% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_3_dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra}
\usage{
dijkstra(graph_v1_v2_w, init_node)
}
\arguments{
\item{graph_v1_v2_w}{A data frame.}

\item{init_node}{source.node from data fram.}
}
\value{
The shortest path (d) of every other node from init_node(s) as a vector \code{x} and \code{y}.
}
\description{
(a,b)=(b,r) if a=b*q+r and r is remaining
}
\details{
Calculate the shortest path (d) of every other node from init_node(s) as a vector by Dijkstra algorithm
}
\seealso{
\url{https://en.wikipedia.org/wiki/dijkstra_algorithm}
}
