\name{Samira_first_package_lab3-package}
\alias{Samira_first_package_lab3-package}
\alias{Samira_first_package_lab3}
\docType{package}
\title{
\packageTitle{Samira_first_package_lab3}
}
\description{
\packageDescription{Samira_first_package_lab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Samira_first_package_lab3}
\packageIndices{Samira_first_package_lab3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Samira_first_package_lab3}

Maintainer: \packageMaintainer{Samira_first_package_lab3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
